+++ apis/cluster/account/v1alpha1/zz_generated.managedlist.go
@@ -44,6 +44,15 @@ func (l *MemberList) GetItems() []resource.Managed {
 	return items
 }
 
+// GetItems of this OAuthClientList.
+func (l *OAuthClientList) GetItems() []resource.Managed {
+	items := make([]resource.Managed, len(l.Items))
+	for i := range l.Items {
+		items[i] = &l.Items[i]
+	}
+	return items
+}
+
 // GetItems of this SubscriptionList.
 func (l *SubscriptionList) GetItems() []resource.Managed {
 	items := make([]resource.Managed, len(l.Items))