+++ apis/cluster/stream/v1alpha1/zz_generated.managedlist.go
@@ -53,6 +53,15 @@ func (l *LiveInputList) GetItems() []resource.Managed {
 	return items
 }
 
+// GetItems of this MOQRelayList.
+func (l *MOQRelayList) GetItems() []resource.Managed {
+	items := make([]resource.Managed, len(l.Items))
+	for i := range l.Items {
+		items[i] = &l.Items[i]
+	}
+	return items
+}
+
 // GetItems of this WatermarkList.
 func (l *WatermarkList) GetItems() []resource.Managed {
 	items := make([]resource.Managed, len(l.Items))